backend/pyproject.toml
[tool.poetry]
name = "gilapi-api"
version = "0.1.0"
description = ""
authors = ["pgmystery <philipp.glaw@gmx.de>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic-settings = "^2.0.3"
motor = "^3.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
httpx = "^0.25.0"
docker = "^6.1.3"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = 'uvicorn src.main:app --reload'
